refactor(stylelint-config): ♻️  Organize plugin `order` config
Resolves issue #85
