+++ packages/stylelint/source/plugins/order/groups/accessibility-interaction.ts
@@ -1,0 +1,28 @@
+import { createPropertiesGroup } from "../index.js";
+
+const group = createPropertiesGroup([
+	"pointer-events",
+	"-ms-touch-action",
+	"touch-action",
+	"cursor",
+	"visibility",
+	"zoom",
+	"table-layout",
+	"empty-cells",
+	"caption-side",
+	"border-spacing",
+	"border-collapse",
+	"content",
+	"quotes",
+	"counter-reset",
+	"counter-increment",
+	"resize",
+	"user-select",
+	"nav-index",
+	"nav-up",
+	"nav-right",
+	"nav-down",
+	"nav-left",
+]);
+
+export default group;
