+++ packages/stylelint/source/plugins/order/groups/background-border.ts
@@ -1,0 +1,65 @@
+import { createPropertiesGroup } from "../index.js";
+
+const group = createPropertiesGroup([
+	"background",
+	"background-color",
+	"background-image",
+	"-ms-filter:\\'progid:DXImageTransform.Microsoft.gradient",
+	"filter:progid:DXImageTransform.Microsoft.gradient",
+	"filter:progid:DXImageTransform.Microsoft.AlphaImageLoader",
+	"filter",
+	"background-repeat",
+	"background-attachment",
+	"background-position",
+	"background-position-x",
+	"background-position-y",
+	"background-clip",
+	"background-origin",
+	"background-size",
+	"background-blend-mode",
+	"isolation",
+	"border",
+	"border-color",
+	"border-style",
+	"border-width",
+	"border-top",
+	"border-top-color",
+	"border-top-style",
+	"border-top-width",
+	"border-right",
+	"border-right-color",
+	"border-right-style",
+	"border-right-width",
+	"border-bottom",
+	"border-bottom-color",
+	"border-bottom-style",
+	"border-bottom-width",
+	"border-left",
+	"border-left-color",
+	"border-left-style",
+	"border-left-width",
+	"border-radius",
+	"border-top-left-radius",
+	"border-top-right-radius",
+	"border-bottom-right-radius",
+	"border-bottom-left-radius",
+	"border-image",
+	"border-image-source",
+	"border-image-slice",
+	"border-image-width",
+	"border-image-outset",
+	"border-image-repeat",
+	"outline",
+	"outline-width",
+	"outline-style",
+	"outline-color",
+	"outline-offset",
+	"box-shadow",
+	"mix-blend-mode",
+	"filter:progid:DXImageTransform.Microsoft.Alpha(Opacity",
+	"-ms-filter:\\'progid:DXImageTransform.Microsoft.Alpha",
+	"opacity",
+	"-ms-interpolation-mode",
+]);
+
+export default group;
