+++ packages/stylelint/source/plugins/order/groups/box-model.ts
@@ -1,0 +1,43 @@
+import { createPropertiesGroup } from "../index.js";
+
+const group = createPropertiesGroup([
+	"box-sizing",
+	"float",
+	"width",
+	"min-width",
+	"max-width",
+	"height",
+	"min-height",
+	"max-height",
+	"padding",
+	"padding-block",
+	"padding-inline",
+	"padding-top",
+	"padding-right",
+	"padding-bottom",
+	"padding-left",
+	"margin",
+	"margin-block",
+	"margin-inline",
+	"margin-top",
+	"margin-right",
+	"margin-bottom",
+	"margin-left",
+	"overflow",
+	"overflow-x",
+	"overflow-y",
+	"-webkit-overflow-scrolling",
+	"-ms-overflow-x",
+	"-ms-overflow-y",
+	"-ms-overflow-style",
+	"overscroll-behavior",
+	"overscroll-behavior-x",
+	"overscroll-behavior-y",
+	"overscroll-behavior-inline",
+	"overscroll-behavior-block",
+	"clip",
+	"clip-path",
+	"clear",
+]);
+
+export default group;
