+++ packages/stylelint/source/plugins/order/groups/flex-grid-children.ts
@@ -1,0 +1,21 @@
+import { createPropertiesGroup } from "../index.js";
+
+const group = createPropertiesGroup([
+	// Flex
+	"flex",
+	"flex-basis",
+	"flex-grow",
+	"flex-shrink",
+	// Grid
+	"grid-area",
+	"grid-row-start",
+	"grid-row-end",
+	"grid-column-start",
+	"grid-column-end",
+	// Shared
+	"align-self",
+	"justify-self",
+	"place-self",
+]);
+
+export default group;
