+++ packages/stylelint/source/plugins/order/groups/position.ts
@@ -1,0 +1,15 @@
+import { createPropertiesGroup } from "../index.js";
+
+const group = createPropertiesGroup([
+	"position",
+	"inset",
+	"inset-block",
+	"inset-inline",
+	"top",
+	"right",
+	"bottom",
+	"left",
+	"z-index",
+]);
+
+export default group;
