+++ packages/stylelint/source/plugins/order/groups/svg.ts
@@ -1,0 +1,40 @@
+import { createPropertiesGroup } from "../index.js";
+
+const group = createPropertiesGroup([
+	"alignment-baseline",
+	"baseline-shift",
+	"dominant-baseline",
+	"text-anchor",
+	"word-spacing",
+	"writing-mode",
+
+	"fill",
+	"fill-opacity",
+	"fill-rule",
+	"stroke",
+	"stroke-dasharray",
+	"stroke-dashoffset",
+	"stroke-linecap",
+	"stroke-linejoin",
+	"stroke-miterlimit",
+	"stroke-opacity",
+	"stroke-width",
+
+	"color-interpolation",
+	"color-interpolation-filters",
+	"color-profile",
+	"color-rendering",
+	"flood-color",
+	"flood-opacity",
+	"image-rendering",
+	"lighting-color",
+	"marker-start",
+	"marker-mid",
+	"marker-end",
+	"mask",
+	"shape-rendering",
+	"stop-color",
+	"stop-opacity",
+]);
+
+export default group;
