+++ packages/stylelint/source/plugins/order/groups/transition-animation.ts
@@ -1,0 +1,21 @@
+import { createPropertiesGroup } from "../index.js";
+
+const group = createPropertiesGroup([
+	"transition",
+	"transition-delay",
+	"transition-timing-function",
+	"transition-duration",
+	"transition-property",
+	"transform",
+	"transform-origin",
+	"animation",
+	"animation-name",
+	"animation-duration",
+	"animation-play-state",
+	"animation-timing-function",
+	"animation-delay",
+	"animation-iteration-count",
+	"animation-direction",
+]);
+
+export default group;
