+++ packages/stylelint/source/plugins/order/groups/typography.ts
@@ -1,0 +1,69 @@
+import { createPropertiesGroup } from "../index.js";
+
+const group = createPropertiesGroup([
+	"font",
+	"font-family",
+	"font-size",
+	"font-style",
+	"font-weight",
+	"font-feature-settings",
+	"font-kerning",
+	"font-variant",
+	"font-variant-ligatures",
+	"font-variant-caps",
+	"font-variant-alternates",
+	"font-variant-numeric",
+	"font-variant-east-asian",
+	"font-variant-position",
+	"font-size-adjust",
+	"font-stretch",
+	"font-effect",
+	"font-emphasize",
+	"font-emphasize-position",
+	"font-emphasize-style",
+	"-webkit-font-smoothing",
+	"-moz-osx-font-smoothing",
+	"font-smooth",
+	"hyphens",
+	"line-height",
+	"color",
+	"text-align",
+	"text-align-last",
+	"text-emphasis",
+	"text-emphasis-color",
+	"text-emphasis-style",
+	"text-emphasis-position",
+	"text-decoration",
+	"text-decoration-line",
+	"text-decoration-thickness",
+	"text-decoration-style",
+	"text-decoration-color",
+	"text-underline-position",
+	"text-underline-offset",
+	"text-indent",
+	"text-justify",
+	"text-outline",
+	"-ms-text-overflow",
+	"text-overflow",
+	"text-overflow-ellipsis",
+	"text-overflow-mode",
+	"text-shadow",
+	"text-transform",
+	"text-wrap",
+	"-webkit-text-size-adjust",
+	"-ms-text-size-adjust",
+	"letter-spacing",
+	"word-break",
+	"word-spacing",
+	"word-wrap", // Legacy name for `overflow-wrap`
+	"overflow-wrap",
+	"tab-size",
+	"white-space",
+	"vertical-align",
+	"list-style",
+	"list-style-position",
+	"list-style-type",
+	"list-style-image",
+]);
+
+export default group;
